+++ tests/test_tfe.py
@@ -62,7 +62,7 @@ class TFEQNodeTests(BaseTest):
         @qml.qnode(self.dev2, interface='tfe')
         def qf(x):
             qml.RX(x, wires=[0])
-            return qml.expval.PauliZ(0), 0.3
+            return qml.expval(qml.PauliZ(0)), 0.3
 
         with self.assertRaisesRegex(QuantumFunctionError, 'must return either'):
             qf(par)
@@ -71,27 +71,27 @@ class TFEQNodeTests(BaseTest):
         @qml.qnode(self.dev2, interface='tfe')
         def qf(x):
             qml.RX(x, wires=[0])
-            ex = qml.expval.PauliZ(1)
-            return qml.expval.PauliZ(0)
+            ex = qml.expval(qml.PauliZ(1))
+            return qml.expval(qml.PauliZ(0))
 
-        with self.assertRaisesRegex(QuantumFunctionError, 'All measured expectation values'):
+        with self.assertRaisesRegex(QuantumFunctionError, 'All measured observables'):
             qf(par)
 
         # ...in the correct order
         @qml.qnode(self.dev2, interface='tfe')
         def qf(x):
             qml.RX(x, wires=[0])
-            ex = qml.expval.PauliZ(1)
-            return qml.expval.PauliZ(0), ex
+            ex = qml.expval(qml.PauliZ(1))
+            return qml.expval(qml.PauliZ(0)), ex
 
-        with self.assertRaisesRegex(QuantumFunctionError, 'All measured expectation values'):
+        with self.assertRaisesRegex(QuantumFunctionError, 'All measured observables'):
             qf(par)
 
         # gates must precede EVs
         @qml.qnode(self.dev2, interface='tfe')
         def qf(x):
             qml.RX(x, wires=[0])
-            ev = qml.expval.PauliZ(1)
+            ev = qml.expval(qml.PauliZ(1))
             qml.RY(0.5, wires=[0])
             return ev
 
@@ -103,7 +103,7 @@ class TFEQNodeTests(BaseTest):
         def qf(x):
             qml.RX(x, wires=[0])
             qml.CNOT(wires=[0, 1])
-            return qml.expval.PauliZ(0), qml.expval.PauliZ(1), qml.expval.PauliX(0)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1)), qml.expval(qml.PauliX(0))
 
         with self.assertRaisesRegex(QuantumFunctionError, 'can only be measured once'):
             qf(par)
@@ -113,7 +113,7 @@ class TFEQNodeTests(BaseTest):
         def qf(x):
             qml.RX(x, wires=[0])
             qml.CNOT(wires=[0, 2])
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         with self.assertRaisesRegex(QuantumFunctionError, 'applied to invalid wire'):
             qf(par)
@@ -123,7 +123,7 @@ class TFEQNodeTests(BaseTest):
         def qf(x):
             qml.RX(x, wires=[0])
             qml.Displacement(0.5, 0, wires=[0])
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         with self.assertRaisesRegex(QuantumFunctionError, 'Continuous and discrete'):
             qf(par)
@@ -132,7 +132,7 @@ class TFEQNodeTests(BaseTest):
         @qml.qnode(self.dev1, interface='tfe')
         def qf(x):
             qml.Displacement(0.5, 0, wires=[0])
-            return qml.expval.X(0)
+            return qml.expval(qml.X(0))
 
         with self.assertRaisesRegex(DeviceError, 'Gate [a-zA-Z]+ not supported on device'):
             qf(par)
@@ -140,9 +140,9 @@ class TFEQNodeTests(BaseTest):
         # ...nor observables
         @qml.qnode(self.dev1, interface='tfe')
         def qf(x):
-            return qml.expval.X(0)
+            return qml.expval(qml.X(0))
 
-        with self.assertRaisesRegex(DeviceError, 'Expectation [a-zA-Z]+ not supported on device'):
+        with self.assertRaisesRegex(DeviceError, 'Observable [a-zA-Z]+ not supported on device'):
             qf(par)
 
     def test_qnode_fanout(self):
@@ -154,7 +154,7 @@ class TFEQNodeTests(BaseTest):
             qml.RX(reused_param, wires=[0])
             qml.RZ(other_param, wires=[0])
             qml.RX(reused_param, wires=[0])
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         thetas = tf.linspace(-2*np.pi, 2*np.pi, 7)
 
@@ -177,20 +177,20 @@ class TFEQNodeTests(BaseTest):
         def circuit_n1s(dummy1, array, dummy2):
             qml.RY(0.5 * array[0,1], wires=0)
             qml.RY(-0.5 * array[1,1], wires=0)
-            return qml.expval.PauliX(0)  # returns a scalar
+            return qml.expval(qml.PauliX(0))  # returns a scalar
 
         @qml.qnode(self.dev1, interface='tfe')
         def circuit_n1v(dummy1, array, dummy2):
             qml.RY(0.5 * array[0,1], wires=0)
             qml.RY(-0.5 * array[1,1], wires=0)
-            return qml.expval.PauliX(0),  # note the comma, returns a 1-vector
+            return qml.expval(qml.PauliX(0)),  # note the comma, returns a 1-vector
 
         @qml.qnode(self.dev2, interface='tfe')
         def circuit_nn(dummy1, array, dummy2):
             qml.RY(0.5 * array[0,1], wires=0)
             qml.RY(-0.5 * array[1,1], wires=0)
             qml.RY(array[1,0], wires=1)
-            return qml.expval.PauliX(0), qml.expval.PauliX(1)  # returns a 2-vector
+            return qml.expval(qml.PauliX(0)), qml.expval(qml.PauliX(1))  # returns a 2-vector
 
         grad_target = (np.array(1.), np.array([[0.5,  0.43879, 0], [0, -0.43879, 0]]), np.array(-0.4))
         cost_target = 1.03257
@@ -222,7 +222,7 @@ class TFEQNodeTests(BaseTest):
             qml.QubitStateVector(np.array([1, 0, 1, 1])/np.sqrt(3), wires=[0, 1])
             qml.Rot(x, y, z, wires=0)
             qml.CNOT(wires=[0, 1])
-            return qml.expval.PauliZ(0), qml.expval.PauliY(1)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliY(1))
 
         @qml.qnode(self.dev2, interface='tfe')
         def circuit1(x, y, z):
@@ -255,7 +255,7 @@ class TFEQNodeTests(BaseTest):
             qml.CNOT(wires=[0, 1])
             qml.RY(y, wires=[0])
             qml.RX(z, wires=[0])
-            return qml.expval.PauliY(0), qml.expval.PauliZ(1)
+            return qml.expval(qml.PauliY(0)), qml.expval(qml.PauliZ(1))
 
         res = circuit(a, b, c)
 
@@ -274,7 +274,7 @@ class TFEQNodeTests(BaseTest):
         def circuit(w, x=None, y=None):
             qml.RX(x, wires=[0])
             qml.RX(y, wires=[1])
-            return qml.expval.PauliZ(0), qml.expval.PauliZ(1)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
 
         circuit = qml.QNode(circuit, self.dev2).to_tfe()
 
@@ -288,7 +288,7 @@ class TFEQNodeTests(BaseTest):
         def circuit(w, x=None):
             qml.RX(x[0], wires=[0])
             qml.RX(x[1], wires=[1])
-            return qml.expval.PauliZ(0), qml.expval.PauliZ(1)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
 
         circuit = qml.QNode(circuit, self.dev2).to_tfe()
 
@@ -303,7 +303,7 @@ class TFEQNodeTests(BaseTest):
 
         def circuit(x, q=default_q):
             qml.RY(x, wires=0)
-            return qml.expval.PauliZ(q)
+            return qml.expval(qml.PauliZ(q))
 
         circuit = qml.QNode(circuit, self.dev2).to_tfe()
 
@@ -319,7 +319,7 @@ class TFEQNodeTests(BaseTest):
 
         def circuit(w, x=None):
             qml.RX(x, wires=[0])
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         circuit = qml.QNode(circuit, self.dev1).to_tfe()
 
@@ -334,7 +334,7 @@ class TFEQNodeTests(BaseTest):
         def circuit(w, x, y):
             qml.RX(x, wires=[0])
             qml.RX(y, wires=[1])
-            return qml.expval.PauliZ(0), qml.expval.PauliZ(1)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
 
         c = circuit(tf.constant(1.), np.pi, np.pi).numpy()
         self.assertAllAlmostEqual(c, [-1., -1.], delta=self.tol)
@@ -346,7 +346,7 @@ class TFEQNodeTests(BaseTest):
         def circuit(w, x=None):
             qml.RX(w, wires=[0])
             qml.RX(x, wires=[1])
-            return qml.expval.PauliZ(0), qml.expval.PauliZ(1)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
 
         circuit = qml.QNode(circuit, self.dev2).to_tfe()
 
@@ -361,7 +361,7 @@ class TFEQNodeTests(BaseTest):
         def circuit(w, x=None):
             qml.RX(w, wires=[0])
             qml.RX(x, wires=[1])
-            return qml.expval.PauliZ(0), qml.expval.PauliZ(1)
+            return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
 
         circuit = qml.QNode(circuit, self.dev2).to_tfe()
 
@@ -379,7 +379,7 @@ class TFEQNodeTests(BaseTest):
             qml.BasisState(input_state, wires=[0, 1])
             qml.RX(x, wires=[0])
             qml.RY(y, wires=[0])
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         circuit = qml.QNode(circuit, self.dev2).to_tfe()
 
@@ -431,7 +431,7 @@ class IntegrationTests(BaseTest):
             qml.RY(phi[1], wires=1)
             qml.CNOT(wires=[0, 1])
             qml.PhaseShift(theta[0], wires=0)
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         @qml.qnode(dev, interface='tfe')
         def circuit_tfe(phi, theta):
@@ -439,7 +439,7 @@ class IntegrationTests(BaseTest):
             qml.RY(phi[1], wires=1)
             qml.CNOT(wires=[0, 1])
             qml.PhaseShift(theta[0], wires=0)
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         phi = [0.5, 0.1]
         theta = [0.2]
@@ -463,7 +463,7 @@ class IntegrationTests(BaseTest):
             qml.RY(phi[1], wires=1)
             qml.CNOT(wires=[0, 1])
             qml.PhaseShift(theta[0], wires=0)
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         @qml.qnode(dev, interface='tfe')
         def circuit_tfe(phi, theta):
@@ -471,7 +471,7 @@ class IntegrationTests(BaseTest):
             qml.RY(phi[1], wires=1)
             qml.CNOT(wires=[0, 1])
             qml.PhaseShift(theta[0], wires=0)
-            return qml.expval.PauliZ(0)
+            return qml.expval(qml.PauliZ(0))
 
         phi = [0.5, 0.1]
         theta = [0.2]